ENH: refactor _intersect_two_rois
